Fix compiler warnings
* include/libwget.h.in (wget_bar_init, wget_bar_update,
wget_bar_print): Change types to int and off_t.
* libwget/bar.c (wget_bar_init, wget_bar_update,
wget_bar_print): Change types to int and off_t.
* src/wget.c (_prepare_file): Add prototype.  Change type of rc to
size_t.

+++ src/bar.c
@@ -99,7 +99,7 @@ void bar_printf(int slotpos, const char *fmt, ...)
 	va_end(args);
 }
 
-void bar_update(int slotpos, size_t max, size_t cur)
+void bar_update(int slotpos, off_t max, off_t cur)
 {
 	wget_thread_mutex_lock(&mutex);
 	wget_bar_update(bar, slotpos, max, cur);
